requirements.txt
Brotli==1.1.0
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
dohq-artifactory==0.9.2
filelock==3.13.1
gitdb==4.0.11
GitPython==3.1.41
identify==2.5.33
idna==3.6
inflate64==1.0.0
Jinja2==3.1.3
MarkupSafe==2.1.4
multivolumefile==0.2.3
nodeenv==1.8.0
numpy==1.26.0
pre-commit==3.6.0
psutil==5.9.7
py7zr==0.20.8
pyarrow==13.0.0
pybcj==1.0.2
pycryptodomex==3.19.0
PyJWT==2.8.0
PyOpenGL==3.1.7
pyppmd==1.1.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyzstd==0.15.9
qt-material==2.14
requests==2.31.0
smmap==5.0.1
texttable==1.7.0
urllib3==2.1.0
virtualenv==20.25.0
